Fix in the nesterov method as well

+++ code/lib/batchtrainer.py
@@ -57,9 +57,9 @@ class BatchTrainer(object):
 
   def buildUpdatesNesterov(self, error, trainingOptions, momentum):
     if trainingOptions.momentumFactorForLearningRate:
-      lrFactor = 1.0 - momentum
+      lrFactor =  np.float32(1.0) - momentum
     else:
-      lrFactor = 1.0
+      lrFactor = np.float32(1.0)
 
     preDeltaUpdates = []
     for param, oldUpdate in zip(self.params, self.oldUpdates):
